==> NASA_telomeres/01a NASA Astros Telos Data Extraction.ipynb <==
# enables access to directories/files
import os

# general purpose modules for handling data
import numpy as np
from numpy import array
import pandas as pd
from pandas import ExcelWriter
from pandas import ExcelFile

# custom module for handling telomere length data
import telomere_methods_astros as telo_ma

---
&nbsp; 

The telomere length measurements by telomere FISH are stored in individual excel sheets 
per timepoint per individual. Thus, we have roughly 200 excel sheets to extract data from. 
Data will be extracted into a dict from which a dataframe is created and saved as a .csv for later use.

&nbsp;

---

# Extracting telomere length data by FISH for all astronauts into a dict

# ls '../data/raw data/'

# this function pulls individual telomere length data from the excel sheets containing
# ImageJ telometer data. DAPI, missing values, and values greater than 
# 3 standard devs from the mean are purged. all telo measurements
# are standardized according to microscope imaged w/, using Cy3 bead values
# to control for microscope intensity differences

dict_astro_individ_telos_dfs = telo_ma.generate_dictionary_for_telomere_length_data(
    '../data/raw data/astros telo data/')


## Making a dataframe from the dict

# takes data from above dict, standardizes # of telomeres according to theoretical max
# for these samples (184 per metaphase, 30 metaphases each sample) using
# random sampling from data per sample
# makes dataframe

astro_df = telo_ma.make_astronaut_dataframe(dict_astro_individ_telos_dfs)
astro_df.head(4)

## Saving all astros telo dataframe to csv for later retrieval

copy_astro_df = astro_df
copy_astro_df['telo data'] = copy_astro_df['telo data'].apply(lambda row: row.tolist())

copy_astro_df.to_csv('../data/compiled data/All_astronauts_telomere_length_dataframe.csv', index = False)

## Reading in astronaut telomere length data per cell

import importlib
%load_ext autoreload
%autoreload 2
%reload_ext autoreload

dict_astro_cell_telos_dfs = telo_ma.grab_astro_telo_values_per_cell_generate_dictionary(
    '../data/raw data/astros telo data/')

astro_cells_df = telo_ma.make_astronaut_cell_data_dataframe(dict_astro_cell_telos_dfs)

astro_cells_df['telo data per cell'] = astro_cells_df['telo data per cell'].apply(lambda row: row.tolist())
astro_cells_df.drop(['Q1', 'Q2-3', 'Q4'], axis=1, inplace=True)

astro_cells_df.to_csv('../data/compiled data/All_astronauts_telomere_length_per_cell_dataframe.csv', index = False)

# Extracting telomere length data by FISH for all astro controls into a dict

dict_ctrl_individ_telos_dfs = telo_ma.grab_control_values_generate_dictionary('../data/raw data/controls telo data')

## Making dataframe from all controls dict

# same as above

control_df = telo_ma.make_control_cell_data_dataframe(dict_ctrl_individ_telos_dfs)
control_df['flight status'] = 'Controls'
control_df.head(2)

## Saving all astro controls telo data for later retrieval

copy_control_df = control_df
copy_control_df['telo data'] = copy_control_df['telo data'].apply(lambda row: row.tolist())
copy_control_df.to_csv('../data/compiled data/All_astro_controls_telomere_length_dataframe.csv', index=False)

## Reading in control telomere length data per cell, saving to file

dict_control_cell_telos_dfs = telo_ma.grab_control_telo_values_per_cell_generate_dictionary(
    '../data/raw data/controls telo data/')

ctrl_cells_df = telo_ma.make_control_cell_data_dataframe(dict_control_cell_telos_dfs)
ctrl_cells_df['telo data per cell'] = ctrl_cells_df['telo data per cell'].apply(lambda row: row.tolist())

ctrl_cells_df.to_csv('../data/compiled data/All_controls_telomere_length_per_cell_dataframe.csv', index = False)

# Reading in astronaut urine/blood biochemical data

biochem_data = pd.read_excel('../data/raw data/Biochemistry means_SD n11 Telomere subjects.xlsx')
biochem_data.shape

## Processing urine biochem data (n=11)

urine_biochem_data = biochem_data.iloc[0:4, 0:11]
urine_biochem_data.rename(columns={'Urine Biochemistry (n=11, not all time points have all 11 subjects though)': 'biochemistry'}, inplace=True)
urine_biochem_data['sample type'] = 'urine'
urine_biochem_data

## Processing blood biochem data (n=11)

blood_biochem_data = pd.read_excel('../data/raw data/Biochemistry means_SD n11 Telomere subjects.xlsx', skiprows=5)
blood_biochem_data = blood_biochem_data.iloc[:, 0:9]
blood_biochem_data.rename(columns={'Blood Biochemistry  (n=11, not all time points have all 11 subjects though)': 'biochemistry'}, inplace=True)
blood_biochem_data['sample type'] = 'blood'
blood_biochem_data.head(4)

## Saving the biochem data to csv's

urine_biochem_data.to_csv('../data/compiled data/urine_biochem_data.csv', index=False)
blood_biochem_data.to_csv('../data/compiled data/blood_biochem_data.csv', index=False)
