==> NASA_telomeres/02a NASA Astros Telo Data Transformation.ipynb <==
# general purpose modules for handling data
import numpy as np
from numpy import array
import pandas as pd

# for loading telo data column containing individual
# telomere length values
from ast import literal_eval

# custom module for handling telomere length data
import telomere_methods_astros as telo_ma

import importlib
%load_ext autoreload
%autoreload 2
%reload_ext autoreload

---
&nbsp; 

The dataframe containing our telomere length data already contains the means. However, we're interested in analyzing (graphing, stats) the individual telomere length measurements. To do so we must extract the values from the list, while keeping the values associated with their timepont/individual. We'll achieve this by 'exploding' the cell containing individual telomere length measurements as a list, into one row per measurement (5520 rows per sample, 184 telos per 30 metaphases per sample). 

As well, we're interested in defining and examining the prevalence of short/long telomeres; this is achieved by finding the values that divide the data into bottom 25% (short telos), middle 50%, and top 25% (long telos). More on that soon.
 
&nbsp;
___

# Transforming telomere length data (teloFISH)

astro_df = pd.read_csv('../data/compiled data/All_astronauts_telomere_length_dataframe.csv')

# literal eval enables interpretation of individual telomere length values in the list 
astro_df['telo data'] = astro_df['telo data'].apply(lambda row: np.array(literal_eval(row)))

astro_df.head(4)

## Exploding list containing individual telomere length measurements into rows

this action explodes the list into columns bearing each datapoint.
picture the list, which contains the individual values, expanding to the
right, up to 5520 measurements

importantly, the index #s per row still refer to the timepont/sample
so we can merge these columns back to the original astro_df
then we'll melt the columns into one, resulting in a lot of rows
where each has one individual telomere length measurement

explode_telos_raw = astro_df['telo data'].apply(pd.Series)

explode_telos_raw.head(2)

exploded_telos_astro_df = (explode_telos_raw
    # merge 5520 columns with original dataframe
    .merge(astro_df, right_index = True, left_index = True)
                           
    # drop unnecessary columns
    .drop(['telo data', 'Q1', 'Q2-3', 'Q4'], axis = 1)
                           
    #specify which columns remain constant per indviidual telo 
    .melt(id_vars = ['astro number', 'astro id', 'timepoint', 'flight status', 'telo means'], value_name = "telo data exploded") 
    .drop("variable", axis = 1)
    .dropna())

exploded_telos_astro_df.shape

copy_exploded_telos_astros_df = exploded_telos_astro_df.copy()

copy_exploded_telos_astros_df.to_csv('../data/compiled and processed data/exploded_telos_astros_df.csv', index = False)

## Loading astro telomere length data per cell, exploding cell data, saving to file

astro_cells_df = pd.read_csv('../data/compiled data/All_astronauts_telomere_length_per_cell_dataframe.csv')

astro_cells_df['telo data per cell'] = astro_cells_df['telo data per cell'].apply(lambda row: np.array(literal_eval(row)))
astro_cells_df['telo data per cell'][0].shape

explode_cells_raw = astro_cells_df['telo data per cell'].apply(pd.Series)

exploded_cells_astro_df = (explode_cells_raw

    .merge(astro_cells_df, right_index = True, left_index = True)
    .drop(['telo data per cell'], axis = 1)
    .melt(id_vars = ['astro number', 'astro id', 'timepoint', 'flight status', 'telo means'], value_name = "telo data per cell") 
    .drop("variable", axis = 1)
    .dropna())

# exploded_cells_astro_df[
#     (exploded_cells_astro_df['astro id'] == 5163) &
#     (exploded_cells_astro_df['timepoint'] == 'L-270')].shape

exploded_cells_astro_df.to_csv('../data/compiled and processed data/exploded_cells_astros_df.csv', index = False)

## Loading astro controls telomere length data per cell, exploding cell data, saving to file

ctrl_cells_df = pd.read_csv('../data/compiled data/All_controls_telomere_length_per_cell_dataframe.csv')

ctrl_cells_df['telo data per cell'] = ctrl_cells_df['telo data per cell'].apply(lambda row: np.array(literal_eval(row)))
ctrl_cells_df['telo data per cell'][0].shape

explode_cells_ctrl_raw = ctrl_cells_df['telo data per cell'].apply(pd.Series)
ctrl_cells_df.head(4)

exploded_cells_ctrl_df = (explode_cells_ctrl_raw

    .merge(ctrl_cells_df, right_index = True, left_index = True)
    .drop(['telo data per cell'], axis = 1)
    .melt(id_vars = ['control id', 'timepoint', 'flight status controls', 'telo means'], value_name = "telo data per cell") 
    .drop("variable", axis = 1)
    .dropna())

# exploded_cells_ctrl_df

exploded_cells_ctrl_df['flight status'] = 'Controls'

exploded_cells_ctrl_df.to_csv('../data/compiled and processed data/exploded_cells_ctrl_df.csv', index = False)

## Defining short/mid/long telomeres w/ quartiles and counting per timepoint per astro

This function identifies the timepoint per astronaut most distal to spaceflight (L-270 or L-180) and identifies the individual telomere length values which divide the data into the bottom 25% (short telos), mid 50%, and top 25% (long telos). The function then counts how many telos reside in each quartile. Now we have a way to define short/long telomeres.

The function then applies those quartile cutoff values to subsequent datapoints and counts how many telomeres reside within each quartile. In doing so, we count the number of telomeres moving into/out of the quartiles, per timepoint per astronaut, and can quantitatively dicuss # of short/long telos.

astro_df.head()

quartiles_astro_df = telo_ma.make_quartiles_columns(astro_df)

for col in ['Q1', 'Q2-3', 'Q4']:
    quartiles_astro_df[col] = quartiles_astro_df[col].astype('int64')

quartiles_astro_df[quartiles_astro_df['astro id'] == 2171]

melted_quartiles_astro_df = pd.melt(
    quartiles_astro_df,
    id_vars = [col for col in quartiles_astro_df.columns if col != 'Q1' and col != 'Q2-3' and col != 'Q4'],
    var_name='relative Q',
    value_name='Q freq counts')

melted_quartiles_astro_df['Q freq counts'] = melted_quartiles_astro_df['Q freq counts'].astype('int64')
melted_quartiles_astro_df['astro id'] = melted_quartiles_astro_df['astro id'].astype('str')

melted_quartiles_astro_df['astro id'].unique()

copy_melted_quartiles_astro_df = melted_quartiles_astro_df
copy_melted_quartiles_astro_df['telo data'] = copy_melted_quartiles_astro_df['telo data'].apply(lambda row: row.tolist())

melted_quartiles_astro_df.to_csv('../data/compiled and processed data/melted_quartiles_astro_df.csv', index = False)

# Cleaning astronaut urine biochem data (n=11)

urine_biochem_data = pd.read_csv('../data/compiled data/urine_biochem_data.csv')
urine_biochem_data

melt_urine_biochem_data = pd.melt(
    urine_biochem_data,
    id_vars=['biochemistry', 'sample type'],
    var_name='timepoint',
    value_name='measurement'
)

melt_urine_biochem_data.head(4)

def make_flight_status_via_timepoint(row):
    if 'Pre' in row:
        return 'Pre-Flight'
    elif 'L' in row:
        return 'Pre-Flight'
    elif 'FD' in row:
        return 'Mid-Flight'
    elif 'R' in row:
        return 'Post-Flight'

def grab_number_remove_plusminus(row):
    if '±' in str(row):
        row = row.split('±')[0].strip()
        row = float(row)
        return row  

melt_urine_biochem_data['flight status'] = melt_urine_biochem_data['timepoint'].apply(lambda row: make_flight_status_via_timepoint(row))

melt_urine_biochem_data['measured analyte'] = melt_urine_biochem_data['measurement'].apply(lambda row: grab_number_remove_plusminus(row))
melt_urine_biochem_data.head(4)

# Reading in astronaut blood biochem data (n=11)

blood_biochem_data = pd.read_csv('../data/compiled data/blood_biochem_data.csv')
blood_biochem_data.head(6)

blood_biochem_data.replace('*', np.NaN, inplace=True)

melt_blood_biochem_data = pd.melt(
    blood_biochem_data,
    id_vars=['biochemistry', 'sample type'],
    var_name='timepoint',
    value_name='measurement')

melt_blood_biochem_data['flight status'] = melt_blood_biochem_data['timepoint'].apply(lambda row: make_flight_status_via_timepoint(row))
melt_blood_biochem_data['measured analyte'] = melt_blood_biochem_data['measurement'].apply(lambda row: grab_number_remove_plusminus(row))
melt_blood_biochem_data.head(4)

melt_blood_biochem_data.rename(columns={'biochemistry':'biochemistry analyte'}, inplace=True)
melt_urine_biochem_data.rename(columns={'biochemistry':'biochemistry analyte'}, inplace=True)

melt_urine_biochem_data.to_csv('../data/compiled and processed data/melt_urine_biochem_data.csv', index=False)
melt_blood_biochem_data.to_csv('../data/compiled and processed data/melt_blood_biochem_data.csv', index=False)
